requirements.txt
## requirement packages
sklearn
pandas
scikit-learn
numpy
streamlit
streamlit-option-menu
Pillow